Demo: some CGI (Python and C) and a lighttpd.conf file for a hands-on example
It's a simple demo for showing how to create the Cookie, since there is
not much info out there. I took the plone Python implementation as a
base but it needs modifications in order to work. The C-code is an
imitation of the mod_authn_tkt.c code since that's the one that will
actually validate the cookies.

github: Closes #2